requirements.txt
# Development and Testing
pre-commit
pytest
pytest-cov
pytest-asyncio
mongomock

# Web Framework
fastapi
uvicorn
typer
requests

# Data Validation and Models
pydantic

# Database and ORM
sqlalchemy
mysql-connector-python
psycopg2-binary
pymongo
keyring
motor

# Data Processing
pandas
numpy
dask
pyarrow
sqlmodel

# Async Support
asyncio
aiofiles

# Utilities
python-dotenv
click
rich
psutil
xlrd
PyJWT

# Type Checking
mypy
types-requests
httpx
openpyxl
apscheduler
